% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeLCMS.R
\name{readMzML}
\alias{readMzML}
\title{Standardize ms1 data to consistent mz/lc values}
\usage{
readMzML(file_name, mz_fidelity_amu = 0.01, lc_fidelity_sec = 1,
  get_ms1features = F, verbose = T)
}
\arguments{
\item{file_name}{path/fileName.mzml}

\item{mz_fidelity_amu}{The sampling fidelity for mz axis. Default 0.01.}

\item{lc_fidelity_sec}{The sampling fidelity for lc axis. Default 1.}

\item{get_ms1features}{Populate the object with XCMS fetaures. Default F.}

\item{verbose}{Print progress. Default T}
}
\description{
Standardize ms1 data to consistent mz/lc values
}
\examples{
dat <- readMzML("example.mzML")
str(dat)
List of 4
$ scans      :List of 1
..$ data:'data.frame':	11075 obs. of  4 variables:
.. ..$ scan_type  : num [1:11075] 1 1 1 1 1 1 1 1 1 1 ...
.. ..$ scan_type_n: int [1:11075] 1 2 3 4 5 6 7 8 9 10 ...
.. ..$ scan_id    : int [1:11075] 1 2 3 4 5 6 7 8 9 10 ...
.. ..$ scan_rt_sec: num [1:11075] 0.35 0.76 1.17 1.59 2 2.41 2.82 3.24 3.66 4.08 ...
$ targets    :List of 1
..$ data:'data.frame':	4773 obs. of  10 variables:
.. ..$ msn_n      : int [1:4773] 1 2 3 4 5 6 7 8 9 10 ...
.. ..$ msn_id     : int [1:4773] 231 264 265 266 267 289 295 307 310 313 ...
.. ..$ pre_mz     : num [1:4773] 470 391 397 436 559 ...
.. ..$ pre_rt_sec : num [1:4773] 102 117 117 117 118 ...
.. ..$ pre_int    : num [1:4773] 0 0 0 0 0 0 0 0 0 0 ...
.. ..$ pre_z      : int [1:4773] 0 0 0 0 0 0 0 0 0 0 ...
.. ..$ pre_ce     : num [1:4773] 0 0 0 0 0 0 0 0 0 0 ...
.. ..$ msn_int_max: logi [1:4773] NA NA NA NA NA NA ...
.. ..$ msn_sn     : logi [1:4773] NA NA NA NA NA NA ...
.. ..$ msn_qual   : logi [1:4773] NA NA NA NA NA NA ...
$ spectra    :List of 2
..$ ms1:List of 3
.. ..$ lc : num [1:7000] 0 0.514 1.029 1.543 2.057 ...
.. ..$ mz : num [1:90000] 350 350 350 350 350 ...
.. ..$ int: int [1:7000, 1:90000] 0 0 0 0 0 0 0 0 16 16 ...
..$ ms2:List of 2
.. ..$ mz : num [1:86000] 70 70 70 70 70 ...
.. ..$ int: int [1:4773, 1:86000] NA NA NA NA NA NA NA NA NA NA ...
$ ms1features: list()
}
\keyword{lcms,}
\keyword{rasterize}
