% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveToHDF5.R
\name{saveToHDF5}
\alias{saveToHDF5}
\title{Save standardize ms1 data to an HDF5 file}
\usage{
saveToHDF5(obj, file_name)
}
\arguments{
\item{obj}{mzh5::readMzML object}

\item{file_name}{The path/filename to save to.}
}
\description{
Save standardize ms1 data to an HDF5 file
}
\examples{
saveToHDF5
}
\keyword{HDF5}
\keyword{save,}
