% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveToRDS.R
\name{saveToRDS}
\alias{saveToRDS}
\title{Save standardize ms1 data to an RDS file}
\usage{
saveToRDS(obj, file_name)
}
\arguments{
\item{obj}{mzh5::readMzML object}

\item{file_name}{The path/filename to save to.}
}
\description{
Save standardize ms1 data to an RDS file
}
\examples{
saveToRDS
}
\keyword{HDF5}
\keyword{save,}
