% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precursorIntensity.R
\name{estPreInt}
\alias{estPreInt}
\title{Standardize ms1 data to consistent mz/lc values}
\usage{
estPreInt(dat, verbose = T)
}
\arguments{
\item{verbose}{Print progress. Default T}

\item{obj}{mzh5::readMzML object}
}
\description{
Standardize ms1 data to consistent mz/lc values
}
\examples{
dat <- readMzML("example.mzML")
dat <- estPreInt(dat)
}
\keyword{intensity}
\keyword{precursor}
